backend/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contextlib2==21.6.0
cron-descriptor==1.4.5
cryptography==43.0.3
distro==1.9.0
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-ninja==1.3.0
django-ninja-extra==0.21.4
django-ninja-jwt==5.3.4
django-redis==5.4.0
django-timezone-field==7.0
flower==2.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
idna==3.10
injector==0.22.0
jiter==0.7.0
kombu==5.4.2
openai==1.53.1
packaging==24.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tornado==6.4.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
